New folder (2)/requirements.txt
langgraph>=0.2.30
langchain>=0.2.16
langchain-openai>=0.1.22
pydantic>=2.8
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
httpx>=0.27.0
fastapi>=0.112.0
uvicorn>=0.30.6
jinja2>=3.1.4

